Refer to Maven Resolver rather than Aether
Closes gh-29255
