+++ spring-boot-project/spring-boot-cli/src/main/java/org/springframework/boot/cli/compiler/grape/RepositorySystemSessionAutoConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2019 the original author or authors.
+ * Copyright 2012-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,8 +20,8 @@ import org.eclipse.aether.DefaultRepositorySystemSession;
 import org.eclipse.aether.RepositorySystem;
 
 /**
- * Strategy that can be used to apply some auto-configuration during the installation of
- * an {@link AetherGrapeEngine}.
+ * Strategy that can be used to apply some auto-configuration during the installation of a
+ * {@link MavenResolverGrapeEngine}.
  *
  * @author Andy Wilkinson
  * @since 1.0.0
